Abstra-Templates/Customer-Onboarding/requirements.txt
slack-sdk==3.23.0
python-dotenv==1.0.0